refactor: use common js files
